Checklists are posted to backend application

+++ app/src/pages/overview/overview.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams } from 'ionic-angular';
+import { Http, Headers, RequestOptions } from '@angular/http';
 
 import { ChecklistArrivalPage } from '../checklist-arrival/checklist-arrival';
 import { ChecklistDeparturePage } from '../checklist-departure/checklist-departure';
@@ -35,7 +36,7 @@ export class OverviewPage {
   checklistLocomotive: ChecklistLocomotive;
   readyForSubmit: boolean = false;
 
-  constructor(public navCtrl: NavController, public navParams: NavParams, private userService: User, private checklistService: Checklist) {
+  constructor(public navCtrl: NavController, public navParams: NavParams, private userService: User, private checklistService: Checklist, public http: Http) {
     this.checklistArrival = new ChecklistArrival();
     this.checklistDeparture = new ChecklistDeparture();
     this.checklistLocomotive = new ChecklistLocomotive();
@@ -71,18 +72,62 @@ export class OverviewPage {
     });
   }
 
-  submitChecklists() {
-    this.checklistService.deleteChecklist("checklistArrival").then((val) => {
-      this.checklistArrival = new ChecklistArrival();
-    });
-    this.checklistService.deleteChecklist("checklistDeparture").then((val) => {
-      this.checklistDeparture = new ChecklistDeparture();
-    });
-    this.checklistService.deleteChecklist("checklistLocomotive").then((val) => {
-      this.checklistLocomotive = new ChecklistLocomotive();
-    });
-    this.readyForSubmit = false;
+  postChecklist(checklistType, checklist) {
+    let headers = new Headers({ 'Content-Type': 'application/json', 'Authorization': 'Bearer GokgwJghPyUoIEfkUZT09fxAbOuvL2Ng' });
+    let options = new RequestOptions({ headers: headers });
+    let postParams = {
+      active: 1,
+      type: checklistType,
+      content: JSON.stringify(checklist),
+    };
+
+    return this.http.post('http://paulpoot.eu/dwg/api/1/tables/checklist/rows', postParams, options);
   }
+
+  submitChecklists() {
+    this.readyForSubmit = false;
+
+    if (this.checklistArrival.saved) {
+      this.postChecklist('Arrival', this.checklistArrival).subscribe(data => {
+        if (data.ok) {
+          this.checklistService.deleteChecklist("checklistArrival").then((val) => {
+            this.checklistArrival = new ChecklistArrival();
+          });
+        } else {
+          console.log(data);
+          this.readyForSubmit = true;
+        }
+      })
+    }
+
+    if (this.checklistDeparture.saved) {
+      this.postChecklist('Departure', this.checklistDeparture).subscribe(data => {
+        if (data.ok) {
+          this.checklistService.deleteChecklist("checklistDeparture").then((val) => {
+            this.checklistDeparture = new ChecklistDeparture();
+          });
+        } else {
+          console.log(data);
+          this.readyForSubmit = true;
+        }
+      })
+    }
+
+    if (this.checklistLocomotive.saved) {
+      this.postChecklist('Locomotive', this.checklistLocomotive).subscribe(data => {
+        if (data.ok) {
+          this.checklistService.deleteChecklist("checklistLocomotive").then((val) => {
+            this.checklistLocomotive = new ChecklistLocomotive();
+          });
+        } else {
+          console.log(data);
+          this.readyForSubmit = true;
+        }
+      })
+    }
+  }
+
+
 
   openChecklist(checklist) {
     this.navCtrl.push(checklist);
